# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Documentation Field Reservation System",
    "description": "by Rofif Fairuz Hawary\n\nLast Updated: 23/04/2024 00:03:XX",
    "contact": {
      "email": "rofifhawary@gmail.com"
    },
    "version": "1.7.7"
  },
...
  "servers": [
    {
      "url": "http://rofif.my.id/"
    },
    {
      "url": "http://localhost:5000/"
    }
  ],
...
  "paths": {
...
    "/player/auth/register": {
      "post": {
        "tags": [
          "player"
        ],
        "summary": "Register user Player",
        "description": "Register user Player",
        "operationId": "registerPlayer",
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerRegisterBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/player/auth/login": {
      "post": {
        "tags": [
          "player"
        ],
        "summary": "Login user Player",
        "description": "Login user Player",
        "operationId": "loginPlayer",
        "requestBody": {
          "description": "Login user Player",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlayerLoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/player/auth/logout": {
      "delete": {
        "tags": [
          "player"
        ],
        "summary": "Logout session user player on used device",
        "description": "Would delete row in Player_Login_Token on device that used",
        "operationId": "logoutPlayer",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerLogoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/admin/auth/register": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Register user Admin",
        "description": "Register user Admin",
        "operationId": "registerAdmin",
        "requestBody": {
          "description": "Created user object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRegisterBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/admin/auth/login": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Login user Admin",
        "description": "Login user Admin",
        "operationId": "loginAdmin",
        "requestBody": {
          "description": "Login user Admin",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLoginBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
...
    "/admin/auth/logout": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Logout session user admin on used device",
        "description": "Would delete row in Admin_Login_Token on device that used",
        "operationId": "logoutAdmin",
        "parameters": [
          {
            "name": "token",
            "in": "header",
            "required": true,
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLogoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "invalid input"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PlayerRegisterBody": {
        "required": [
          "username",
          "password",
          "name",
          "phone"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "devexample"
          },
          "password": {
            "type": "string",
            "example": "examplepass"
          },
          "name": {
            "type": "string",
            "example": "Dev Example"
          },
          "phone": {
            "type": "string",
            "example": "08123456789"
          }
        },
        "xml": {
          "name": "PlayerRegisterBody"
        }
      },
      "PlayerRegisterResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "register_status": {
            "type": "boolean"
          }
        },
        "xml": {
          "name": "PlayerRegisterResponse"
        }
      },
      "PlayerLoginBody": {
        "required": [
          "username",
          "password",
          "virtual_device_id"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "devexample"
          },
          "password": {
            "type": "string",
            "example": "devexample"
          },
          "virtual_device_id": {
            "type": "string",
            "example": "some_device_id"
          }
        },
        "xml": {
          "name": "PlayerLoginBody"
        }
      },
      "PlayerLoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "login_status": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "ava_url": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              }
            }
          }
        },
        "xml": {
          "name": "PlayerLoginResponse"
        }
      },
...
      "PlayerLogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "logout_status": {
            "type": "boolean"
          }
        },
        "xml": {
          "name": "PlayerLogoutResponse"
        }
      },
      "AdminRegisterBody": {
        "required": [
          "username",
          "password",
          "name",
          "phone"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "devexample"
          },
          "password": {
            "type": "string",
            "example": "examplepass"
          },
          "name": {
            "type": "string",
            "example": "Dev Example"
          },
          "phone": {
            "type": "string",
            "example": "08123456789"
          }
        },
        "xml": {
          "name": "AdminRegisterBody"
        }
      },
      "AdminRegisterResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "register_status": {
            "type": "boolean"
          }
        },
        "xml": {
          "name": "AdminRegisterResponse"
        }
      },
      "AdminLoginBody": {
        "required": [
          "username",
          "password",
          "virtual_device_id"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "devexample"
          },
          "password": {
            "type": "string",
            "example": "devexample"
          },
          "virtual_device_id": {
            "type": "string",
            "example": "some_device_id"
          }
        },
        "xml": {
          "name": "AdminLoginBody"
        }
      },
      "AdminLoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "login_status": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "ava_url": {
                "type": "string"
              },
              "username": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "phone": {
                "type": "string"
              }
            }
          }
        },
        "xml": {
          "name": "AdminLoginResponse"
        }
      },
      "AdminLogoutResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "logout_status": {
            "type": "boolean"
          }
        },
        "xml": {
          "name": "AdminLogoutResponse"
        }
      },
...
    }
  }
}